+++ app/Http/Controllers/Summary/View/ResourceController.php
@@ -1,53 +1,52 @@
 <?php
 
-namespace App\Http\Controllers\Summary;
+namespace App\Http\Controllers\Summary\View;
 
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 use App\HttpResponse\Header;
-use App\Models\Summary\Subcategory;
-use App\HttpOptionResponse\SummarySubcategoryCollection;
+use App\Models\Summary\Resource;
+use App\HttpOptionResponse\SummaryResourceCollection;
 use App\HttpRequest\Parameter;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Support\Facades\Config;
 
 /**
- * Summary controller for the subcategories routes
+ * Summary controller for the resource routes
  *
  * @author Dean Blackborough <dean@g3d-development.com>
  * @copyright Dean Blackborough 2018-2022
  * @license https://github.com/costs-to-expect/api/blob/master/LICENSE
  */
-class SubcategoryView extends Controller
+class ResourceController extends Controller
 {
     /**
-     * Return a summary of the subcategories
+     * Return a summary of the resources
      *
-     * @param $resource_type_id
-     * @param $category_id
+     * @param string $resource_type_id
      *
      * @return JsonResponse
      */
-    public function index(Request $request, $resource_type_id, $category_id): JsonResponse
+    public function index(Request $request, string $resource_type_id): JsonResponse
     {
         if ($this->hasViewAccessToResourceType((int) $resource_type_id) === false) {
-            return \App\HttpResponse\Response::notFoundOrNotAccessible(trans('entities.category'));
+            return \App\HttpResponse\Response::notFoundOrNotAccessible(trans('entities.resource-type'));
         }
 
         $cache_control = new \App\Cache\Control($this->user_id);
-        $cache_control->setTtlOneMonth();
+        $cache_control->setTtlOneWeek();
 
         $cache_summary = new \App\Cache\Response\Summary();
         $cache_summary->setFromCache($cache_control->getByKey($request->getRequestUri()));
 
         if ($cache_control->isRequestCacheable() === false || $cache_summary->valid() === false) {
             $search_parameters = Parameter\Search::fetch(
-                Config::get('api.subcategory.summary-searchable')
+                Config::get('api.resource.summary-searchable')
             );
 
-            $summary = (new Subcategory())->totalCount(
+            $summary = (new Resource())->totalCount(
                 $resource_type_id,
-                $category_id,
+                $this->viewable_resource_types,
                 $search_parameters
             );
 
@@ -62,7 +61,7 @@ class SubcategoryView extends Controller
             }
 
             $collection = [
-                'subcategories' => $total
+                'resources' => $total
             ];
 
             $headers = new Header();
@@ -84,20 +83,19 @@ class SubcategoryView extends Controller
 
 
     /**
-     * Generate the OPTIONS request for the subcategories summary
+     * Generate the OPTIONS request for the resource summary
      *
-     * @param $resource_type_id
-     * @param $category_id
+     * @param string $resource_type_id
      *
      * @return JsonResponse
      */
-    public function optionsIndex($resource_type_id, $category_id): JsonResponse
+    public function optionsIndex(string $resource_type_id): JsonResponse
     {
         if ($this->hasViewAccessToResourceType((int) $resource_type_id) === false) {
-            return \App\HttpResponse\Response::notFoundOrNotAccessible(trans('entities.category'));
+            return \App\HttpResponse\Response::notFoundOrNotAccessible(trans('entities.resource-type'));
         }
 
-        $response = new SummarySubcategoryCollection($this->permissions((int) $resource_type_id));
+        $response = new SummaryResourceCollection($this->permissions((int) $resource_type_id));
 
         return $response->create()->response();
     }
